# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the Mando server API documentation",
        "title": "Mando API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/skills/add": {
            "post": {
                "description": "Add one or more skills",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skills"
                ],
                "summary": "Add new skills",
                "parameters": [
                    {
                        "description": "Skills to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skills.AddSkillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Skills added successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "skills.AddSkillRequest": {
            "description": "Model for adding skills",
            "type": "object",
            "required": [
                "skills"
            ],
            "properties": {
                "skills": {
                    "description": "List of skills to add",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Go"
                    ]
                }
            }
        }
    }
}